refactor: extract image actions from AgentWorkspace
