+++ src/components/agent/session-submission-types.ts
@@ -1,7 +1,24 @@
 import type * as React from "react";
+
+export type SubmitHermesSession = (
+  content: string,
+  explicitSession?: HermesSessionInfo,
+  options?: {
+    issueReport?: PendingIssueReport;
+    displayContent?: string;
+    titleContent?: string;
+    attachments?: AgentAttachment[];
+    selectSession?: boolean;
+    onAttachmentsUpdated?: (attachments: AgentAttachment[]) => void;
+    modelTarget?: CapturedSessionModelTarget;
+    dispatchReservation?: HermesSessionDispatchReservation;
+    skipPrompt?: boolean;
+  },
+) => Promise<string | undefined>;
 import type { AgentProjectContext } from "../../lib/agent-project-context";
 import { ProjectContextSignatureStore } from "../../lib/agent-project-context";
 import { HermesGatewayClient } from "../../lib/hermes-gateway";
+import type { HermesSessionDispatchReservation } from "../../lib/hermes-session-dispatch-mutex";
 import { type SessionModelSelectionMap } from "../../lib/hermes-session-model-selection";
 import {
   type HermesSessionInfo,
